Port Preview app to Lit; guard content URLs by published state
- platforms-preview: drop React entirely; fold the preview-load logic
  (subscription code, previewHtml POST, ETag dedupe, iframe injection,
  10s refresh) into a single UmbLitElement. Interval cleared/restarted from
  the unique observer + authFetchChanged + connectedCallback, cleared in
  disconnectedCallback; in-flight loads guarded by a #disconnected flag.
  Removed react/react-dom peerDeps + @types, react:false, non-React tsconfig.
- GetContentUrlsHandler: only emit staging/production URLs for published
  content (restores the v13 SetUrl Published-state gate that the migration
  had dropped).
