+++ src/Cloud/N3O.Umbraco.Cloud.Platforms/frontend/platforms-preview/package.json
@@ -7,14 +7,8 @@
         "build": "tsc --noEmit && vite build",
         "watch": "vite build --watch"
     },
-    "peerDependencies": {
-        "react": "^19.0.0",
-        "react-dom": "^19.0.0"
-    },
     "devDependencies": {
         "@repo/build-config": "*",
-        "@n3oltd/backoffice-core": "*",
-        "@types/react": "^19.0.0",
-        "@types/react-dom": "^19.0.0"
+        "@n3oltd/backoffice-core": "*"
     }
 }
